release version 2.0.2

+++ src/attach.cpp
@@ -13,24 +13,25 @@
  *  $Id$
  */
 
+#include "compiler.h"
 #ifdef HAVE_CONFIG_H
-  # include "aconfig.h"
+    #include "aconfig.h"
 #endif
 
 #include <sys/types.h>
 #include <sys/stat.h>
 #ifdef HAVE_UNISTD_H
-  #include <unistd.h>
+    #include <unistd.h>
 #endif
 #ifdef HAVE_DOS_H
-  #include <dos.h>
+    #include <dos.h>
 #endif
 #include <stdio.h>
 #include <stdlib.h>
 #include "string.hpp"
 #include <ctype.h>
 #ifdef HAVE_FCNTL_H
-  #include <fcntl.h>
+    #include <fcntl.h>
 #endif
 #include <errno.h>
 #include "constant.hpp"
@@ -45,17 +46,17 @@
 #include "mytypes.hpp"
 #include "msg.hpp"
 #if defined(HAS_IO_H) || defined(HAVE_IO_H)
-  #include <io.h>
+    #include <io.h>
 #endif
 
 #ifdef __WATCOMC__
-  #undef far
+    #undef far
 #endif
 
 #ifdef __NT__
-  #undef byte
-  #undef EXPENTRY
-  #include <windows.h>
+    #undef byte
+    #undef EXPENTRY
+    #include <windows.h>
 #endif
 
 // --------------------------------------------------------------------
@@ -117,7 +118,7 @@ static char * FullName(char * name)
 
     if((strchr(name, PATHDELIMC) == NULL
 #ifndef __unix__
-        && strchr(name, ':') == NULL
+            && strchr(name, ':') == NULL
 #endif
        ) || (IgnoreAttachPath != FALSE))
     {
@@ -158,7 +159,7 @@ int _AttToSize(char * Buff)
         if(i != ENOENT)
         {
             Log.Level(LOGE) << "   Unable to open file: '" << tmt <<
-                               "', Errno: " << i << EOL;
+                            "', Errno: " << i << EOL;
             return FALSE;
         }
         else
@@ -174,7 +175,7 @@ int _AttToSize(char * Buff)
         i = errno;
         close(fh);
         Log.Level(LOGE) << "   Unable to get size of file: '" << tmt <<
-                           "', Errno: " << i << EOL;
+                        "', Errno: " << i << EOL;
         return FALSE;
     }
 
@@ -326,7 +327,7 @@ int _AttachExists(char * Buff)
     if(SkipHiddenFiles)
     {
         struct _finddata_t fd;
-        int h = _findfirst(tmt, &fd);
+        int h = (int)_findfirst(tmt, &fd);
 
         if(h != -1)
         {
@@ -396,7 +397,7 @@ int _MoveAttach(char * Buff)
         CharToOem(tmt, tmt);
 #endif
         Log.Level(LOGE) << "   Unable to move file '" << tmt << "' to " <<
-                           NewPath << EOL;
+                        NewPath << EOL;
         return FALSE;
     }
 
@@ -423,7 +424,7 @@ int _CopyAttach(char * Buff)
         CharToOem(tmt, tmt);
 #endif
         Log.Level(LOGI) << "   File not found: '" << tmt <<
-                           "', can't copy." << EOL;
+                        "', can't copy." << EOL;
         return TRUE;
     }
 
@@ -441,7 +442,7 @@ int _CopyAttach(char * Buff)
         CharToOem(tmt, tmt);
 #endif
         Log.Level(LOGE) << "   Unable to copy file '" << tmt << "' to " <<
-                           NewPath << EOL;
+                        NewPath << EOL;
         return FALSE;
     }
