+++ src/utils.cpp
@@ -41,12 +41,12 @@
 #else
     #include <malloc.h>
 #endif
+#include "compiler.h"
 #include "log.hpp"
 #include "mytypes.hpp"
 #include "constant.hpp"
 #include "utils.hpp"
 #include "vars.hpp"
-#include <smapi/compiler.h>
 #include "wildmat.hpp"
 
 #ifdef __WATCOMC__
@@ -366,7 +366,7 @@ void nls_strupr(char * s)
 
     if(s2)
     {
-        int slen = strlen(s);
+        int slen = (int)strlen(s);
 
         OemToChar(s2, s);
         LCMapString(LOCALE_SYSTEM_DEFAULT, LCMAP_UPPERCASE, s, slen, s2,
@@ -444,7 +444,7 @@ const char * dirslashbug(const char * dirname)
 //   len = GetFullPathName(dirname, MAX_PATH, newname, &p);
 //   if ((len > 3) && (newname[len-1] == PATHDELIMC)) {
     strcpy(newname, dirname);
-    len = strlen(newname);
+    len = (int)strlen(newname);
 
     if((len > 1) && (newname[len - 1] == PATHDELIMC))
     {
@@ -460,7 +460,11 @@ const char * dirslashbug(const char * dirname)
  */
 unsigned short & FirstWord(const unsigned int & Dword)
 {
-    struct SplitDWORD {unsigned short first; unsigned short second;};
+    struct SplitDWORD
+    {
+        unsigned short first;
+        unsigned short second;
+    };
     SplitDWORD * p = (SplitDWORD *)&Dword;
     return p->first;
 }
@@ -481,11 +485,11 @@ int fsCompareName(const char * Name, const char * Mask)
 }
 
 #if defined (__WATCOMC__) && defined (MSDOS)
-    extern "C" 
-    {
+extern "C"
+{
     // special function. Replaces located in SMAPI.
-        void pascal far flush_handle2_(int fh){}
-    }
+    void pascal far flush_handle2_(int fh) {}
+}
 #endif
 
 int SetSemaphore(void)
@@ -510,7 +514,7 @@ int SetSemaphore(void)
         if(SemTime == SemaphoreTime)
         {
             Log.Level(101) << "Detected semaphore file '" << Semaphore <<
-                              "'." << EOL;
+                           "'." << EOL;
         }
 
         sleep(1);
@@ -523,12 +527,12 @@ int SetSemaphore(void)
         if(access(Semaphore, F_OK) != 0)
         {
             Log.Level(101) << "Semaphore file '" << Semaphore <<
-                              "' now released." << EOL;
+                           "' now released." << EOL;
         }
         else
         {
             Log.Level(101) << "Semaphore file '" << Semaphore <<
-                              "' now ignored by time." << EOL;
+                           "' now ignored by time." << EOL;
         }
     }
 
@@ -538,7 +542,7 @@ int SetSemaphore(void)
     if(fh == NULL)
     {
         Log.Level(100) << "Unable to create semaphore file '" << Semaphore <<
-                          "'." << EOL;
+                       "'." << EOL;
         return FALSE;
     }
 
@@ -557,7 +561,7 @@ int ReleaseSemaphore(void)
     return TRUE;
 }
 
-extern "C" 
+extern "C"
 {
     void ShowLogLine(char * msg)
     {
